issue 2622 refactor
